chore(thirdFeedback): rename serverside test files to sanitise codebase - add CONTRIBUTING.md - red-design README.md - update README.md - update api-docs
