+++ package.json
@@ -5,7 +5,7 @@
   "main": "server.js",
   "scripts": {
     "test": "npm run seq:db && nyc mocha server/test/ --timeout 25000 --require babel-register --exit",
-    "test:env": "npm run seq:db && nyc npm test",
+    "test:env": "cross-env NODE_ENV=test npm run seq:db && nyc mocha server/test/ --timeout 25000 --require babel-register --exit",
     "test:tester": "cross-env NODE_ENV=test npm run seq:db && nyc --reporter=html --reporter=text --reporter=lcov --reporter=text-lcov _mocha server/test/ --compilers js:babel-core/register --exit",
     "start:prod": "NODE_ENV=production babel-node  server/src/server.js",
     "start:dev": "NODE_ENV=development nodemon server/src/server.js --exec babel-node",
@@ -118,6 +118,7 @@
     "style-loader": "^0.19.0",
     "supertest": "^3.0.0",
     "swagger-jsdoc": "^1.9.7",
+    "swagger-ui-express": "^3.0.8",
     "validator": "^9.1.1",
     "webpack": "^3.8.1",
     "webpack-dev-middleware": "^1.12.1",
